Responsive web design fixes
