+++ src/core/api/models/index.ts
@@ -1,4 +1,5 @@
 export * from './auth';
+export * from './links';
 export * from './notifications';
 export * from './patterns';
 export * from './responses';
