+++ src/core/api/models/links.ts
@@ -1,0 +1,4 @@
+export interface Link {
+  label: string;
+  to: string;
+}
