+++ src/modules/home/sections/search/Search.scss
@@ -10,6 +10,11 @@
   p {
     width: 470px;
     text-align: center;
+
+    @include tablet {
+      max-width: 470px;
+      width: auto;
+    }
   }
 
   h1 {
