+++ src/shared/components/footer/Footer.scss
@@ -5,10 +5,19 @@
   grid-template-columns: 1fr 1200px 1fr;
   background: $primaryDark;
 
+  @include laptop {
+    grid-template-columns: 1fr 100fr 1fr;
+  }
+
   .wrapper {
     display: grid;
     grid-template-columns: 1fr 1fr 1fr;
     grid-column: 2/3;
+
+    @include media-max-width(950px) {
+      display: flex;
+      flex-direction: column;
+    }
 
     .description {
       @include labelThin;
@@ -37,6 +46,10 @@
     .feedback {
       padding: 32px 60px 42px 0;
 
+      @include laptop {
+        padding-left: 60px;
+      }
+
       figure {
         @include row;
 
@@ -59,6 +72,13 @@
       padding: 36px 60px 42px 60px;
       border-right: 1px solid $white;
       border-left: 1px solid $white;
+
+      @include media-max-width(950px) {
+        border-right: 0;
+        border-left: 0;
+        border-top: 1px solid $white;
+        border-bottom: 1px solid $white;
+      }
 
       .description {
         margin: 32px 0;
