+++ src/shared/components/navbar/Navbar.tsx
@@ -1,29 +1,22 @@
-import React from 'react';
+import React, { useCallback, useState } from 'react';
 import { NavLink } from 'react-router-dom';
 
 import { Logo } from 'ui';
 
+import { BASE_LINKS, IMPORTANT_LINKS } from './utils';
+
+import { Sidebar } from './sidebar';
+import { SidebarTrigger } from './sidebar-trigger';
+
 import csx from './Navbar.scss';
 
-namespace Navbar {
-  export interface Link {
-    label: string;
-    to: string;
-  }
-}
-
-const BASE_LINKS: Navbar.Link[] = [
-  { label: 'Home', to: '/' },
-  { label: 'About', to: '/about' }
-];
-
-const IMPORTANT_LINKS: Navbar.Link[] = [
-  { label: 'App', to: '/app' },
-  { label: 'Register', to: '/register' },
-  { label: 'Log In', to: '/login' }
-];
-
 const Navbar = () => {
+  const [sidebarOpen, setSidebarOpen] = useState(false);
+
+  const handleSidebarToggle = useCallback(() => {
+    setSidebarOpen((sidebarOpen) => !sidebarOpen);
+  }, []);
+
   return (
     <nav className={csx.navbar}>
       <div className={csx.wrapper}>
@@ -33,7 +26,6 @@ const Navbar = () => {
           </figure>
           <span>Jupi.io</span>
         </div>
-
         <div className={`${csx.links} ${csx.baseLinks}`}>
           {BASE_LINKS.map(({ label, to }) => (
             <NavLink key={to} to={to} activeClassName={csx.activeLink} exact={true}>
@@ -49,7 +41,9 @@ const Navbar = () => {
             </NavLink>
           ))}
         </div>
+        <SidebarTrigger sidebarOpen={sidebarOpen} setSidebarOpen={handleSidebarToggle} />
       </div>
+      {sidebarOpen ? <Sidebar /> : null}
     </nav>
   );
 };
