+++ src/shared/components/navbar/index.ts
@@ -1,1 +1,3 @@
 export { default as Navbar } from './Navbar';
+
+export * from './utils';
