+++ src/shared/components/navbar/sidebar-trigger/SidebarTrigger.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import csx from './SidebarTrigger.scss';
+
+const Sidebar = ({ sidebarOpen, setSidebarOpen }) => {
+  const burgerClass = sidebarOpen ? csx.burgerActive : csx.burgerInactive;
+  return (
+    <button className={`${csx.burger} ${burgerClass}`} onClick={() => setSidebarOpen()}>
+      <div />
+      <div />
+      <div />
+    </button>
+  );
+};
+
+export default Sidebar;
