+++ src/shared/components/navbar/sidebar-trigger/index.ts
@@ -1,0 +1,1 @@
+export { default as SidebarTrigger } from './SidebarTrigger';
