+++ src/shared/components/navbar/sidebar/index.ts
@@ -1,0 +1,1 @@
+export { default as Sidebar } from './Sidebar';
