+++ src/shared/components/navbar/utils.ts
@@ -1,0 +1,12 @@
+import { Link } from 'core/api/models';
+
+export const BASE_LINKS: Link[] = [
+  { label: 'Home', to: '/' },
+  { label: 'About', to: '/about' }
+];
+
+export const IMPORTANT_LINKS: Link[] = [
+  { label: 'App', to: '/app' },
+  { label: 'Register', to: '/register' },
+  { label: 'Log In', to: '/login' }
+];
