+++ src/shared/components/templates-grid/TemplatesGrid.scss
@@ -7,6 +7,18 @@
   grid-template-columns: repeat(4, auto);
   grid-gap: 32px;
 
+  @include media-max-width(1450px) {
+    grid-template-columns: repeat(3, auto);
+  }
+
+  @include media-max-width(1150px) {
+    grid-template-columns: repeat(2, auto);
+  }
+
+  @include media-max-width(750px) {
+    grid-template-columns: repeat(1, auto);
+  }
+
   .spaceholder {
     width: 320px;
     height: 444px;
