+++ src/styles/media-queries.scss
@@ -1,0 +1,39 @@
+$mobile: 425px;
+$tablet: 900px;
+$laptop: 1300px;
+
+@mixin media-max-width($width) {
+  @media (max-width: $width) {
+    @content;
+  }
+}
+
+@mixin media-min-width($width) {
+  @media (min-width: $width) {
+    @content;
+  }
+}
+
+@mixin media-between-width($min-width, $max-width) {
+  @media (min-width: $min-width) and (max-width: $max-width) {
+    @content;
+  }
+}
+
+@mixin mobile() {
+  @include media-max-width($mobile) {
+    @content;
+  }
+}
+
+@mixin tablet() {
+  @include media-max-width($tablet) {
+    @content;
+  }
+}
+
+@mixin laptop() {
+  @include media-max-width($laptop) {
+    @content;
+  }
+}
