+++ src/styles/mixins.scss
@@ -1,4 +1,5 @@
 @import './palette.scss';
+@import './media-queries';
 
 @mixin col {
   display: flex;
